Finished XNAT Datasource refactor
